Update patch set 1
Patch Set 1:

But this implies will implies that label and contentDescription should use the overriden package name (that implies to load the context in builder). I though this more like an internal usage (by the system). Apps should't use in the 99% of the cases, since they are going to have its own resources.

Patch-set: 1