Update patch set 4
Patch Set 4:

Looks good. I feel like somewhere, maybe in the javadoc for SDK_INT, you should specify that if the device is not CM or the CM SDK is not present, it will return 0.

I know your code snippet in this patch will work for that case since you check >=, but it still feels like the unsupported return value is undefined to me as a developer.

Patch-set: 4