Update patch set 7
Patch Set 7:

(1 comment)

Sorry, but which is the idea behind this patch? Why we need it as part of cmsdk?

AFAIK, SubscriptionManager and getActiveSubscriptionInfoList() are public, and part of the public api of Android (via TELEPHONY_SUBSCRIPTION_SERVICE). And one also can subscribe to changes vid OnSubscriptionsChangedListener.

Is there anything i missed?

Patch-set: 7