Update patch set 7
Patch Set 7:

> (1 comment)
 > 
 > Sorry, but which is the idea behind this patch? Why we need it as
 > part of cmsdk?
 > 
 > AFAIK, SubscriptionManager and getActiveSubscriptionInfoList() are
 > public, and part of the public api of Android (via
 > TELEPHONY_SUBSCRIPTION_SERVICE). And one also can subscribe to
 > changes vid OnSubscriptionsChangedListener.
 > 
 > Is there anything i missed?

The idea behind this is to provide access to the dual SIM functionalities. The current patch was intended to check if I was making things right.

I'm currently working on making other function that are not available through Android SDK, and that would enable 3rd party applications to leverage the dual SIM capabilities of a phone.

Patch-set: 7