Update patch set 6
Patch Set 6:

It seems like an error to be able to set both a resource id and a Bitmap as the icon.  It appears from the SystemUI patch that the resource id is given preference.  It may be worth throwing in the Builder if a second icon is trying to be set.

Patch-set: 6