Update patch set 1
Patch Set 1:

But yeah, though you are facing it in SetupWizard, what really should be done is, the device mainatiner should set the file permissions in the init scripts. Because these are device specific things, and if a device is properly maintained, there shouldn't be sysfs permission errors unless you are adding a key in CMHardware and then forgetting to set the permission during init.

Patch-set: 1