Update patch set 1
Patch Set 1:

Realistically "Allows an app to make changes to the airplane mode and mobile data network settings" should change the actual name of the permission itself because the permission constant was created in terms that it could expand to contain a further subset of options for whatever interfacing entity. Since we already shipped the jar to maven, this change won't even appear anywhere else but the docs and a yet-non-shipping 12.1 branch in the future.

Patch-set: 1