Update patch set 1
Patch Set 1:

The reason it was changed was because this will be displayed in the user facing dialog whenever any third party app that requests this permission is being installed on the device. It helps the user to understand exactly what is being installed. I intend to cherry-pick these string changes to the shipping branch, so they would be displayed on the system even for 12.1.

I think it still makes sense to leave the permission names the same because they still could be expanded to cover additional use cases in the future. At that point, we would also modify the string to describe the additional Settings that it covers.

Patch-set: 1