Update patch set 7
Patch Set 7:

I have few comments/questions on this, PerformanceManager.java to be more specific.

First, mProfileSetByUser is unused.
Second, there are some issues with the profiles. Whenever I try to change profile, the profile is set back to "balanced" as soon as the "Battery" setting page is closed. In addition to that, the profile is set to balanced on boot, after it has been correctly set to what is saved in the preferences.

All these problems seem to be solved by simply dropping activityResumed(), which seems to be the cause of these unwanted profile changes.

I haven't read the code in its entirety, so I'm just leaving this comment here.

Patch-set: 7
Reviewer: Gabriele M <13897@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0