Update patch set 3
Patch Set 3:

It is not possible to execute the setZenMode tests without contaminating the test suite.
Pre-test, a zen mode (priority or quiet mode) can be already set indefinitely or without duration. But we have no way of knowing its duration. This is because the CountdownConditionProvider timer does not keep state of countdowns previously submitted. So, remove these tests from this test suite for now. This test is only a unit test for making sure PartnerInterface can be packed/unpacked. Include setAirplaneMode because its state can be restored cleanly.

Patch-set: 3