Update patch set 1
Patch Set 1:

Do we really need a UI element for this setting? If yes then the setting should be configurable if its even shown or not. we should  put this in the dialer/phone app -> settings -> display options or advanced. Keep the default behavior as what it is right now, and then people who are impacted by this issue can go turn off CLOP

Patch-set: 1