Update patch set 1
Patch Set 1:

We throw on write if the key is undefined within CMSettings, this is per the new mechanism in android system settings(can't write whatever you want anymore).

Like Danny mentions, there are other ways (like system properties) to persist this -- however I'm not sure of what the gating qualifier for an option to be included within CMSettings is. Does it need to impact multiple devices? Is that sufficient?

Patch-set: 1