Update patch set 1
Patch Set 1:

Adnan- it looks like something subtle changed recently which has the system calling into CMHW before it's ready. This is a quick workaround, but we really should throw an exception if the service is MIA and crash the client. I'm gone for the day but can look into it later tonight or tomorrow.

Patch-set: 1