Update patch set 1
Patch Set 1:

Should we null out the private instance if the service isn't found?

I don't think this should be reverted wholesale since it also fixes inconsistency. My issue here is that some of these things (like the persistent storage bits) should never silently fail if they are supposed to work because they could have disastrous effects. Regardless if its an error elsewhere we've hit the ordering problem multiple times and it should be trivially obvious what is happening. Right now it is not. 

We drop a feature file for org.cyanogenmod.hardware. Let's crash the manager if the service isn't available and the system feature is found.

Patch-set: 1