Update patch set 1
Patch Set 1:

We shouldn't null it. If the only scenario that is to be worried about is ordering issues by someone attempting to utilize this service before it is up, we can mitigate that by checking against caller being system uid and logging as needed. (as we discussed).

It also shouldn't be the services interfaces responsibility to verify a mandatory service is available (since we're not core to every devices functionality), something like that would be hosted in a validation suite (like parts of CTS).

Changing the backing interface to not be static in a singleton pattern where the entry point is a static method is a bit misleading with this change, and could break the assumptions of any existing consumer which would never hit the ordering issue, since something like that only exists in platform development.

Patch-set: 1