Update patch set 2
Patch Set 2: Code-Review-1

Adnan, I'm sorry but I just don't follow you on why the backing instance needs to be static. The whole point of a singleton is so that you can have non-static state inside of an object which can be retrieved in a static way. I also don't understand your argument against using the system feature to decide how to behave- you're going for "absolute correctness" but not addressing the case where someone made a mistake and didn't understand a subtle implementation detail. Or in this particular case, it was just missed. Had the system just crashed when the input changes were made, this would have been caught instantly instead of manifesting in an unexpected way (broken LiveDisplay).

Patch-set: 2