Update patch set 2
Patch Set 2:

Most app developers won't ever need to keep track of anything. The system service will be up by the time they try to act up on the interface for it(this manager).

There's a secondary patch up to warn on getService() service instance being null because of specific scenarios which may cause it.

It just doesn't make sense to me to crash a client process because a modular system with no mandatory services doesn't have a backend implementation available at a very early point in init.

Patch-set: 2