Update patch set 1
Patch Set 1: Code-Review-1

This should live in the global table. If the use case is for a client to be able to read a read only property, then that is the table that provides it. You can then avoid having to do the IPC transaction in your successive patch to client binder calling identity. The SDK can then just read the value as needed for the client process.

Patch-set: 1