Update patch set 1
Patch Set 1:

The reason the other parceling tests attempt to do each item individually is to make sure the marshalling and unmarshalling can't snag an off by one error. It might be beneficial to sprinkle in some nonsequential members to be written to the parcel and verified when unraveled?

Patch-set: 1