Update patch set 5
Patch Set 5:

This works great and all, the app recovers if killed, but it doesn't fix the issue that nothing wakes it up once it's dead if only callback mode is used. I'm going to add a tiny platform service which uses this to handle sending broadcasts for this stuff and possibly route changes as well.

Patch-set: 5