Update patch set 1
Patch Set 1:

Yeah, something is not right here. 

1. Do you have a cmhw implementation?
2. Does it override DisplayColorCalibration?
3. Does it need to?
4. Does it actually work?

My guess is that somehow the base implementation (which works for every device) is being stomped on by something which doesn't work?

That said, I need to add error checks for this case, but I think your cmhw is messed up and returning false for support of DisplayColorCalibration.

Patch-set: 1