Update patch set 1
Patch Set 1:

Functionally they are exactly the same.. We should definitely use an overlay in f/b to choose the class to launch (not a system property like I used). I'm a little partial to the service-starting-services approach for a couple of reasons..

1. Familiar API and no need for reflection in f/b
2. Central place to publish any sort of management or capabilities/discovery service
3. Service approach allows for getting lifecycle events. I don't have a great usecase for this this though :)

In either case, they are 100% functionally equivalent. No additional threads or anything of that sort end up created.

Patch-set: 1