Update patch set 1
Patch Set 1: Code-Review-1

Have been thinking about this throughly, and IMHO, I believe this logic does not belong to the manager. The temp unit is just a (global) setting, specific to weather, yet a setting which is not controlled by the manager. This logic should be moved to settings

Patch-set: 1
Label: Code-Review=-1