Update patch set 1
Patch Set 1:

(2 comments)

I added the cm api changes required to make it buildable.
If it needs to be in a different commit, np, except this is
the commit introducing those resources so guess it's meant to be here.

Will try now, looks promising.

From screenshots, I agree with LuK1337, would be nice if the logo
could respect the CM logo dimensions on most devices,
and if possible with the transition not too direct,
since Cid would go down in bootanimation, and instantly show up.

Patch-set: 1