Update patch set 2
Patch Set 2:

https://drive.google.com/file/d/0Bw4WlD4fmC5xaWllWXpUdWQyZW8/view?usp=sharing

Initial rev. Versions for both revealing the individual apps as well as obscuring as the regular implementation does. While having a standard, determinate progress bar gets you a quick idea of where you are in the process, the bar here is indeterminate as users should know that activity is being made and the phone has not frozen. This is why the circular indeterminate progress animation is there now.

The branded element is up the top corner and small to maximize black screen area, although if its customizable, the placement and size could change. I just don't know if it's worth playing with the boot animation to blend this in somehow when some (like the current CM anim) have end segments.

Finally, the third screen shows an example of the finalizing state before the lock screen loads in.

The whole design is modeled after a Now Playing state in a music player, where a similar amount of information is shown.

Joey has offered to build this as we have a process for sharing spec in place, but I wanted to post a rev up here as well.

Patch-set: 2