Update patch set 2
Patch Set 2:

https://drive.google.com/file/d/0Bw4WlD4fmC5xbGI1WFd1aXFvbzA/view?usp=sharing

Another small rev. Progress is determinate and an animated ellipsis has been added. My one worry about that is it being difficult to track while the string is updating from app names, but it's not a big thing. I've also subdued the branding element and centered it.

Patch-set: 2