Update patch set 2
Patch Set 2:

It looks better like this. For the "not trackable" part: when apps don't take long to dexopt it will be the change of appname showing there is still work going on. If it's a bigger app, the dots should be trackable good enough. If apps are obscured there won't be a change of text at all so again the ellipsis can be easily spotted. No worries from my side here

Patch-set: 2