Update patch set 4
Patch Set 4:

I tried playing with animations and my opinion is that it's easier for us to bundle a video player and play a Disney movie while the user is waiting.
The "show the user the process is not dead" idea is kinda pointless since we're showing which app is being optimized, I tested this on a msm8974 and the app icon is changed sth like each 2 seconds - so I don't think we need additional "signals of life" here. But, if you can't live without magic animations I can add them.
@danny I've addressed all your comments (except those that were not possible / working). Also I'm not 100% sure I like the idea of having 2 separate classes for "isApk" state and "!isApk"

Patch-set: 4