Update patch set 3
Patch Set 3:

Minimizing footprint is the goal. I want to get CM in a place where we are able to do all the things we do without shitting up AOSP in the process. Having access to the entire codebase kind of makes us spoiled, and when it comes time to merge things it's a nightmare. This time, Google changed all of settings to use the support libraries. All of our stuff needs converted to use it anyway, and this seems like the perfect opportunity to get it done.

It would be really nice if all the menus in Settings had the same kind of functionality as the dashboard and you could just plug stuff in remotely. That was the inspiration for this. Ready for review  and working now too :)

Patch-set: 3