Update patch set 3
Patch Set 3: Code-Review-1

I just feel that this solution goes way overboard. Minimizing footprint is a good goal, but IMHO the additional gains in footprint minimization isn't worth the effort in both code complication and runtime overhead.
Also, we can only replace PreferenceScreens with this, but can't replace ListPreferences, SwitchPreferences and custom ones, which are the ones needing the deepest integration, so this solution, even if it's technically cool, doesn't prevent merge issues in the future.
(Also, conversion from framework preferences to the support library ones really is a no brainer)

Patch-set: 3