Update patch set 3
Patch Set 3:

I have to disagree, considering that settings already uses a fairly complex mechanism for adding dashboard tiles. 

Open to alternatives, but we won't be adding all our stuff to the settings app again. Otherwise I am going to run with this.

To do any other preference types we still need an rpc mechanism, this is just to cover the use case you mentioned. There is no real runtime overhead here, Binder was made for this stuff and all we are doing with this is showing a tile and giving the ability to update it from CMParts.

The goal here is to hide the complexity behind a simple API, which I would say a one liner is pretty good. The preference will also be able to remove itself too. With the logic in the remote app.

Patch-set: 3