Update patch set 3
Patch Set 3:

Danny, this isn't meant to be a competition :)  If you have a better solution which solves the problem, let's do it. As of right now, you are just saying "this sucks" without a better solution.

The current stuff is made for dealing with our preferences, specifically the myriad of PreferenceScreens we add which link out to big piles of code. The other types of prefs are trivial and most can be covered with SystemSettingsSwitchPreference or the CM variants. I'm currently adding support for custom XML attributes which can specify the a preference either requires a feature (cm:requiresFeature) or a package (cm:requiresPackage). This addresses pretty much all the use cases which are left in settings that can't be moved to CMParts and we end up with simple XML edits instead of twisty mazes of code.

Patch-set: 3