Update patch set 3
Patch Set 3:

I don't see this as a competition. TBH, I see this as a solution for a problem that doesn't exist.

I know in repeating myself, but: What actual, real world merge issues are solved by this solution? Adding new code or strings in separate, new files doesn't cause any merge issues. Editing XML files to add preferences might cause them, but one has to do that both with the existing approach as well as with this solution. Editing code to adjust the new prefs might cause merge issues, but one would also need them with this solution (e.g. disabling dependent prefs) or there's an easy alternative solution (e.g. a summary provider callback provided from XML, or a custom ListPreference for the cases where the list contents need to be messed with).

My alternative solution is: just do everything as in cm-13.0, maybe with the exceptions for summaries etc. mentioned above. It doesn't complicate the code base, it doesn't require sprinkling the settings code and strings over 3 repos (which enlarges effort in patch management), and it has the same merge impact as this solution.

Patch-set: 3