Update patch set 4
Patch Set 4:

Having done 12 or so full Android merges, I can say for sure that this reduces merge pain. It's far easier to port over a standalone app than to port a thousand patches into an architecture that's been changing for the last year behind closed doors. Right now, SystemUI and Settings are the most painful parts of the process. Telephony/Dialer is not far behind. We went to great pains to minimize impact on the framework itself with CMSDK. It took me 5 minutes to get all the features that live there working. For M, it took 3 solid guys 2 weeks to get Settings done. CMParts would have been done in a day.

Patch-set: 4