Update patch set 4
Patch Set 4: Code-Review-2

> Having done 12 or so full Android merges, I can say for sure that
 > this reduces merge pain. It's far easier to port over a standalone
 > app than to port a thousand patches into an architecture that's
 > been changing for the last year behind closed doors. Right now,
 > SystemUI and Settings are the most painful parts of the process.
 > Telephony/Dialer is not far behind. We went to great pains to
 > minimize impact on the framework itself with CMSDK. It took me 5
 > minutes to get all the features that live there working. For M, it
 > took 3 solid guys 2 weeks to get Settings done. CMParts would have
 > been done in a day.

No, it wouldn't, because you _still_ need to merge in the XML changes and _still_ need to merge in custom prefs and _still_ need to merge in changes that handle pref dependencies, which are (and I know I'm repeating myself)
a) the actual contention points causing the largest headaches and
b) the cases NOT solved by this approach.

I'm not sure why you're ignoring my questions about those items that I've asked multiple times above and why you're ignoring my ping requests on IRC, but I'm sure I need to see solid answers to them before being convinced that this approach is a good idea. Just playing the 'just trust me' card is not sufficient for that.

Patch-set: 4