Update patch set 4
Patch Set 4:

I'm done then, I have heard no constructive feedback here other than "this sucks". I have a problem to solve here and this is a step in the right direction. You are blocking me because you want things to stay the way they were. If you want to go and re-do a week of work, have at it. Or perhaps we could just iterate instead of blocking?

Patch-set: 4