Update patch set 4
Patch Set 4:

> I'm done then, I have heard no constructive feedback here other
 > than "this sucks". I have a problem to solve here and this is a
 > step in the right direction.

You believe it's a step in the right direction. I believe it's a step in the wrong direction. Why do you think my view can't possibly be correct?

 > You are blocking me because you want
 > things to stay the way they were.

I am blocking this (that is, -2 instead of -1) because I want to discuss the problems this approach induces, and I see no other way of getting you to even _think_ about potential issues this approach has. At the moment, you don't seem to see any.

 > If you want to go and re-do a
 > week of work, have at it. Or perhaps we could just iterate instead
 > of blocking?

Good idea. I asked how you're going to approach the actually problematic items 2 days ago, but you didn't bother to answer yet. As it stands, this patch doesn't solve the problem at hand (which might be worth solving, but I don't see how yet). My proposed solutions of 'use a separate source directory' and 'keep everything as it was' solve the problem equally well as your solution.

And I understand it's frustrating to invest quite a bit of time into a solution people don't like, but that surely can't be a reason to bypass code review, can it?

Patch-set: 4