Update patch set 4
Patch Set 4:

Now we're talking :)

Ideally, I want a solution that eliminates the need for the XML and layout changes in settings too. Like how the settings dashboard is done. I think the big items that live behind a PreferenceScreen should absolutely be in a standalone app. The only reason for the RPC is so they could update their summary views. The alternative to the RPC + callback would be to do what Dashboard does and query PM for the metadata, we would just need a way for the app to tell Settings to re-check the metadata. I don't see a benefit one way or the other, but I think my implementation is simple enough and accomplishes the goals.

Going forward though, perhaps additional categories could be added. Like an component which would list it's metadata as being in the com.android.settings.display category, and it could supply an anchor preference or priority on where it should be placed. I suspect this shouldn't be too hard, and would bring these to zero-impact. 

For listpreferences and other simple inline stuff, the metadata concept could go even further. For a system setting switch, you just need the pref type (switch) and the settings key. For a list, you need the key, the entryvalues and the entries. The settings app would create one of our custom widgets (which would live in CMSDK) for these cases, display it in the category we specified and anchor it to the right place in the list. Anything more complicated should just use a preferencescreen approach TBH. We would need to add something in SettingsPreferenceFragment (which everything is based on) to call into our catalog/lookup code and insert the prefs into the right places. This was my plan for the next steps after making some progress here.

I'm not a huge fan of the PM / manifest based approach since a service like PartsCatalog is easier for us to control and more simple to implement.

Patch-set: 4