Update patch set 4
Patch Set 4:

Re: static linking, you still have to modify a bunch of stuff inside of settings, and your code is also now dependent on it. I prefer a standalone app because it doesn't have to deal with internal refactoring. If we want to refactor it to match, we can, but it should mostly just work.

The ideal result (which isn't gonna happen here today) would be a single patch to Settings which allows us to insert custom preferences anywhere in the hierarchy, defined by an external APK.

Patch-set: 4