Update patch set 4
Patch Set 4:

Honestly, we need to move on here. Gone are the days where it's OK to spew stuff all over the codebase. Android has gotten way too big and complex, and we can either keep re-doing this dance every six months or we can change the way we do things so the pain is reduced. 

For example, with settings, I know you say that the change to the support library is "trivial", but you are actually wrong. We need to fix *every single file* to deal with RecyclerView and the general flow changes. Take a look at commit 39b467482d1bf256a111c757e9b7621c6f523271. We could have completely deferred this if we had a standalone app, and done it a month or two down the road at our leisure.

Patch-set: 4