Update patch set 3
Patch Set 3:

I'm not sure I'd pick this work up or not. It was begun because we wanted the ability to update platform services without having to perform an OTA because of the possibility of shipping on certain carriers. Since this is off the table (at least for now), there is probably better stuff to work on.

That said, I think slicing it into an app-per-service was a bit heavy handed. It might have been better to put the entire backend into a single APK and use the AIDL no-op generator to sort out the stubs.

Patch-set: 3
Reviewer: Steve Kondik <32@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0