Update patch set 2
Patch Set 2:

In my opinion, for 14.1 (and not 13.0), we should kill all the old overlays,
to remove duplication & avoid such issues.

It's the same logic as Enhanced NFC, the worst thing that can happen is a device build breaking & the update is really easy, no potential faulty build possible.

Patch-set: 2