Update patch set 4
Patch Set 4:

The overlay should be set by all devices using CM, in a way it was good it broke,
but broken things are bad so no.

The current state is LIGHTS_PULSATING because this is the only overlay that used to be true by default before,
hence follow the old logic and stay compatible 100% to the old behaviour,
that way devices who never set anything can continue to do so.

Patch-set: 4