performance manager: Honour power profiles after disabling LPM
As described in

BUGBASH-351

In Settings >> "Battery saver and performance", select any Power profile
other than Balanced and then turn on Extreme power saver and then turn
off Extreme power saver (LPM).

Prior to this change the Power profile would be reset to Balanced.  After
this change, it remains what the user had previously selected.

Note: I had to add the extra storeProfile option because changing the
profile and disabling LPM are parallel operations.

That is, if you have LPM enabled and change the profile, LPM is disabled
and the power profile is changed in parallel.  That means you cannot
use LPM as an indication of whether or not the profile needs to be saved
and instead need to provide that knowledge explicitly to the underlying code.

Change-Id: Ie204db753c9002cfd25a7ede297cd1df0cd898ff