Update patch set 2
Patch Set 2:

This is not cm7 where the more toggles you have the better it is. In a real world scenario no one is going to switch a such thing an and off often so it'd just be another "look how cool it is: my rom has more toggles than yours. I can toggle anything feature on and off because it's just cool seeing the entry in the settings list".

Also me (and others) are not fans of changing the aosp behavior, so I'm not even sure if the fw-side change will be ever merged

Patch-set: 2