Update patch set 2
Patch Set 2:

> This is not cm7 where the more toggles you have the better it is.
 > In a real world scenario no one is going to switch a such thing an
 > and off often so it'd just be another "look how cool it is: my rom
 > has more toggles than yours. I can toggle anything feature on and
 > off because it's just cool seeing the entry in the settings list".
 > 
 > Also me (and others) are not fans of changing the aosp behavior, so
 > I'm not even sure if the fw-side change will be ever merged

It's not a question of being better because there are more toggles. Like Dan said, people would accidentally toggle mobile data, that's why I added a setting to enable this feature, which is disabled by default. It's not useless and it would be useful for many users: you don't have to expand the data tile to enable it, you can just one tap and toggle it.

And regarding changing the AOSP behavior, well, if you're not fans of changing it, then why does Lineage have all the features it has today? I'm pretty sure most of them change the AOSP behavior and you're perfectly fine with it.

I know it's your opinion about it, but now I'm telling you mine, and it's not a reasonable argument to use to reject something that you think is going to be useless.

Patch-set: 2
Reviewer: Alexandre Oliveira <15485@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0