cmsdk: Prevent profiles config corruptions
Instead of overwriting the current configuration, rename the old
config, write the new config in a new file and delete the old config
once we are sure the new config has been written. In this way the next
time we will try to load the configuration from disk we will know
whether the last config update was successful or not and in case it
wasn't, load the backup copy. By doing so we will lose the latest
changes the user did, but we will still be able to recover the last
good config.

BUGBASH-1012

Change-Id: I432263e460b52a8f9b69c921e1a531b4fc1b4338