cmsdk: Update ring volume/mode only if needed
DND is automatically enabled during phone calls. This means that
if the current profile overrides the ring volume or mode and the
user has applied different ring volume or mode overriding the
profile overrides, the user changes get lost in favor of the profile
override once the DND gets disabled at the end of the phone call.
Apply the profile overrides only if we couldn't apply them during
the last profile change to fix this.

This should also prevent ring volume changes when profiles are
disabled, but the current profile is not null.

This fixes commit dd66df03d6a0395ef922f35b5116e7b6356a32f0
("cmsdk: Don't change ring volume or mode if DND is enabled")

Change-Id: Iabac796f5d63768822e4f4b67c84f48cfb827456