cmsdk: Update ring volume/mode only if needed
DND is enabled automatically during phone calls and then disabled.
If the current profile overrides the ring volume or mode, we
shouldn't the profile settings if the user has manually changed the
ring volume or mode without changing profile. Apply the profile
overrides only if we couldn't during the last profile change not to
override any user preference.

This also includes the case in which profiles are not configured. A
profile is always selected, even if not active. If we restore the
profile configuration whenever DND gets disabled, we will override
the user preference.

This fixes commit 3d3d2e9c3263be8fb63a49de0d3ddabdac5d5581
("cmsdk: Don't change ring volume or mode if DND is enabled")

REGRESSION-1142

Change-Id: Iabac796f5d63768822e4f4b67c84f48cfb827456