Profiles: Don't force disable the lockscreen
setKeyguardEnabled() is meant for external apps that want to
temporarily disable the lockscreen (e.g. phone app). In addition
to disabling the lockscreen it disables some inputs (e.g. home
button). If a profile hides the losckscreen using setKeyguardEnabled(),
it enables the input restrictions which are removed only once a
different profile re-enables the lockscreen. This can happen if a
profile that disables the lockscreen gets activated by a trigger while
the lockscreen is shown. Don't force disable the lockscreen and wait
for the user to unlock the device not to have the restrictions applied.

BUGBASH-1322

Change-Id: I1cd980b21e3a12a04e791b113a694bb2d9b61bd3