Update patch set 2
Patch Set 2:

> Wolfram, would you be willing to also look at porting tethering
 > stats:
 > https://review.lineageos.org/#/c/LineageOS/android_lineage-sdk/+/213367/
 > 
 > (Disclaimer: I've not looked at whether the same API is available
 > in nougat)

Sam, the API NetworkStats.STATS_PER_UID is missing in Nougat, so probably
no easy porting.
However, the tethering traffic in NOUGAT is not ignored, on the contrary,
it is displayed "doubled".

I tried it two ways (Phone A is gateway, Phone B is final receiver):
* Mobile-data (Phone A) tethered via WiFi to Phone B
* WiFi data (Phone A) tethered via Bluetooth to Phone B.

Both behave the same:
If Phone B has a download of 100 kB/s and an upload of 50 kB/s, then Phone A displays a network traffic of 150 kB/s for upload as well as for download,
i.e. in NOUGAT the traffic of both network interfaces are summed up, although
probably only the "external" network interface (mobile data or WiFi) should be
observed for our NetworkTraffic class.
Is the situation different in OREO?

Patch-set: 2
Reviewer: Wolfram Liebchen <11409@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>